requirements.txt
ddddocr==1.5.6
loguru==0.7.2
parsel==1.8.1
pycryptodomex==3.21.0
Requests==2.32.3
